split services aggregated copy into its original imported files (in extern subdir)
